# Faster and more general reductions for sparse matrices