@IainNZ 2015-03-16
Hah, thats a 1520x speed up for `var(A,2)`, delightful!