@tkelman 2015-03-17
Doesn't this assume the function being mapped is pure? While that's the case for the higher-level wrappers around mapreduce, it's not necessarily true for general mapreduce.