@simonster 2015-03-17
That's true, or at least it assumes that `f(0)` is pure beyond allocation etc. 

We could have a `mapreduce_pure` and use that for all of the wrappers, and then have a separate `mapreduce` for SparseMatrixCSC that calls `f(0)` as many times as there are zeros. Alternatively, we could specify that `mapreduce` expects a pure function (at least for SparseMatrixCSC) and suggest using `mapfoldl`/`mapfoldr` for impure functions.

How common is it that you want to pass an impure function to `mapreduce`? Anything that cares about the order in which it's called is out of the question since the associativity is undefined, but I suppose you could still pass a function that injects randomness or has side effects. OTOH, I don't think I've ever seen code that does this.