@tkelman 2015-03-17
I don't expect impure functions to be common with `mapreduce` at all, but the concern came up last time `map` for sparse matrices was asked about. For `map` I suggested a differently-named function to perform the analogous operation only on the structurally nonzero values, but in the case of `mapreduce` perhaps it's not necessary since there are already implicit restrictions due to associativity.