@simonster 2015-03-17
I noted the possible reuse of return values of `f` in the docs for `mapreduce`. The constraint is not quite as strong as requiring functional purity of `f` since e.g. it's fine if every invocation of `f` returns a new array even though `f` would not be a pure function. If anyone has a better way to word it, I'm open to suggestions.