@ViralBShah 2015-03-19
We should also document mapreduce for sparse.