@simonster 2015-03-19
Does it need separate documentation? The interface is the same as `mapreduce` for dense. I guess I could add this to `NEWS.md`, though.