@ViralBShah 2015-03-20
Adding to NEWS sounds like a good idea - and also the performance improvements deserve a mention in NEWS.